parse journal messages and convert to logstash event format
This update adds parsing of the raw Journal JSON messages into Logstash
V1Events. This means the server-side conversion code can be removed
from the Logstash server config.

Previously in v0.1.x, no parsing of JSON was handled by
journal-2-logstash. We simply shipped the raw JSON directly to the
Logstash server. This meant we needed to use several filters on the
logstash server to convert the raw Journal JSON messages into a
Logstash compatible Event format. This is now handled within
journal-2-logstash so that the server-side filtering in Logstash is no
longer necessary.

There may be performance impacts since we are now unmarshaling and
re-marshaling JSON within the clients. This is likely offset by less
CPU usage on the Logstash server.